refactor(tilemap): :art: Move tilepicker to seperate struct
